remove soft ref caching from model caches
LruCache's ReferenceQueue means that soft references take two collections
to be freed. During the first collection the softreference will be cleared,
freeing the memory for the User, Track or Playlist. The SoftReference
instance itself won't be freed. It will sit in the reference queue
until the next time cleanSoftMap is called. That is potentially horrible
for memory fragmentation problems, since by default we would have 700
small objects that need two passes to be freed.

Also, because containsKey doesn't check the softmap, many calls through
scModelManager won't find items that are in the softmap and they will
be repopulated, instead of being retrieved form cache.

+++ app/src/main/java/com/soundcloud/android/cache/ModelCache.java
@@ -4,11 +4,9 @@ import com.soundcloud.android.model.ScModel;
 
 public class ModelCache<V extends ScModel> extends LruCache<Long, V> {
     /**
-     * 2 level cache - LRU (bound to capacity) + softreference map (unbound)
-     *
      * @param capacity max capacity for the LRU cache
      */
-    public ModelCache(long capacity) {
+    public ModelCache(int capacity) {
         super(capacity);
     }
